Remove .zip file and honor KeepDays in cleanup of BcNuGet cache

+++ ContainerHandling/Flush-ContainerHelperCache.ps1
@@ -175,8 +175,13 @@ try {
     
         if (($caches.Contains('all') -or $caches.Contains('bcnuget')) -and ($bcContainerHelperConfig.BcNuGetCacheFolder) -and (Test-Path $bcContainerHelperConfig.BcNuGetCacheFolder)) {
             Get-ChildItem -Path $bcContainerHelperConfig.BcNuGetCacheFolder | Where-Object { $_.PSIsContainer } | ForEach-Object {
-                Write-Host "Remove $($_.FullName.SubString($bcContainerHelperConfig.BcNuGetCacheFolder.Length+1))"
-                Remove-Item -Path $_.FullName -Recurse -Force
+                Get-ChildItem -Path $_.FullName | ForEach-Object {
+                    $lastWrite = $_.LastWriteTime
+                    if ($keepDays -eq 0 -or $lastWrite -lt (Get-Date).AddDays(-$keepDays)) {
+                        Write-Host "Remove $($_.FullName.SubString($bcContainerHelperConfig.BcNuGetCacheFolder.Length+1))"
+                        Remove-Item -Path $_.FullName -Recurse -Force
+                    }
+                }
             }
         }
